==> 02-DataManipPSSS.Rmd <==
---
title: "02b - DataManipPSSS"
author: "Danielle Ethier"
date: "2022-10-04"
output: html_document
editor_options: 
  chunk_output_type: console
---

Start by loading the packages needed for this chapter

```{r packages, eval=FALSE}

#install.packages("remotes")
#remotes::install_github("BirdsCanada/naturecounts")

require(naturecounts)
require(tidyverse)
require(unmarked)
require(lubridate)
require(reshape)
require(reshape2)
require(nloptr)
require(ggpubr)
require(measurements)


out.dir <- paste("Output/")
```

Data manipulation of the Puget Sound Seabird Survey (PSSS) dataset, provided in October 2022 by Seattle Audubon. 

```{r data}

PSSS<-read.csv("Data/PSSSdata.csv")
BMDE<-meta_bmde_fields(version="core")

```

Compile the complete species list now for use later in the data compilation. The data tables you will need can be accessed using the naturecounts R package.

```{r sp, eval=FALSE}

sp.code<-meta_species_codes()
sp.code<-sp.code %>% filter(authority=="BSCDATA") %>% select(-authority, -species_id2, -rank) %>% distinct()

sp.tax<-meta_species_taxonomy()
sp.tax<-sp.tax %>% select(species_id, scientific_name, english_name) %>% distinct()

sp<-left_join(sp.code, sp.tax, by="species_id")
sp<-sp %>% distinct(english_name, .keep_all = TRUE)

```


PSSS Data Manipulation to put in BMDE framework

```{R data manip}

#break apart position and reform into lat and long

PSSS<-read.csv("Data/PSSSdata.csv")

PSSS$lat<-sub(" W.*", "", PSSS$position)  
PSSS$long<-sub(".*W", "", PSSS$position)

PSSS$lat = gsub('N', '', PSSS$lat)
PSSS$long = gsub('W', '', PSSS$long)

PSSS$DecimalLatitude = measurements::conv_unit(PSSS$lat, from = 'deg_dec_min', to = 'dec_deg')
PSSS$DecimalLatitude<-as.numeric((PSSS$DecimalLatitude))
PSSS$DecimalLongitude = measurements::conv_unit(PSSS$long, from = 'deg_dec_min', to = 'dec_deg')
PSSS$DecimalLongitude<-as.numeric(PSSS$DecimalLongitude)
PSSS$DecimalLongitude=PSSS$DecimalLongitude*(-1)

#break apart survey_date and reform into day, month, year
PSSS<-PSSS %>% separate(survey_date, into=c("Date", "del"), sep=" ") %>% select(-del) %>% separate(Date, into=c("YearCollected", "MonthCollected", "DayCollected"), sep="-") 
#wrangle raptor data into the long format since each species identification should be in a unique row. 
raptor1<-PSSS %>% filter(raptor1 != "") %>% mutate(common_name = raptor1, bird_count = raptor1_count, notes= raptor1_affect)%>%  select(-raptor1, -raptor2, -raptor3, -raptor1_count, -raptor2_count, -raptor3_count, -raptor1_affect, -raptor2_affect, -raptor3_affect) 

raptor1<-raptor1 %>% group_by(site_name, common_name, YearCollected, MonthCollected, DayCollected) %>% mutate(bird_count=sum(bird_count)) %>% distinct(common_name, site_name, YearCollected, MonthCollected, DayCollected, .keep_all=TRUE)

raptor2<-PSSS %>% filter(raptor2 != "") %>% mutate(common_name = raptor2, bird_count = raptor2_count, notes= raptor2_affect)%>%  select(-raptor1, -raptor2, -raptor3, -raptor1_count, -raptor2_count, -raptor3_count, -raptor1_affect, -raptor2_affect, -raptor3_affect) 

raptor2<-raptor2 %>% group_by(site_name, common_name, YearCollected, MonthCollected, DayCollected) %>% mutate(bird_count=sum(bird_count)) %>% distinct(common_name, site_name, YearCollected, MonthCollected, DayCollected, .keep_all=TRUE)

raptor3<-PSSS %>% filter(raptor3 != "") %>% mutate(common_name = raptor3, bird_count = raptor3_count, notes= raptor3_affect) %>%  select(-raptor1, -raptor2, -raptor3, -raptor1_count, -raptor2_count, -raptor3_count, -raptor1_affect, -raptor2_affect, -raptor3_affect) 

raptor3<-raptor3 %>% group_by(site_name, common_name, YearCollected, MonthCollected, DayCollected) %>% mutate(bird_count=sum(bird_count)) %>% distinct(common_name, site_name, YearCollected, MonthCollected, DayCollected, .keep_all=TRUE)

PSSS<-PSSS %>%  select(-raptor1, -raptor2, -raptor3, -raptor1_count, -raptor2_count, -raptor3_count, -raptor1_affect, -raptor2_affect, -raptor3_affect) 

#bind raptor data back with PSSS data
PSSS<-rbind(PSSS, raptor1)
PSSS<-rbind(PSSS, raptor2)
PSSS<-rbind(PSSS, raptor3)

#remove rows with missing common name
PSSS<-PSSS %>% filter(common_name !="")

#remove bearing and distance because we want each species/ site/ date to be a single row in the data set similar to BBCWS

PSSS<-PSSS %>% select(-bearing, -dist)

#Now summarize the records per speices/ sit/ date
PSSS<-PSSS %>% group_by(site_name, common_name, YearCollected, MonthCollected, DayCollected) %>% mutate(bird_count=sum(bird_count)) %>% distinct(common_name, site_name, YearCollected, MonthCollected, DayCollected, .keep_all=TRUE)

#replace Thayer's Gull with Ivory Gull
PSSS<-PSSS %>% mutate(common_name = ifelse(common_name == "Thayer's Gull", "Ivory Gull", common_name))

#Merge with species ID
PSSS<-merge(PSSS, sp, by.x=c("common_name"), by.y= ("english_name"), all.x=TRUE)
  
#rename data columns to match BMDE
PSSS<-PSSS %>% dplyr::rename(CommonName =common_name, SurveyAreaIdentifier= survey_site_id, Locality = site_name, MinimumElevationInMeters=elevation, MaximumElevationInMeters=elevation, TimeObservationsStarted=start_time, TimeCollected = start_time, TimeObservationsEnded=end_time, ObservationCount = bird_count, ObservationCount2=large_flock_best, ObsCountAtLeast = large_flock_min, ObsCountAtMost = large_flock_max, FieldNotes=notes, Collector = name, ScientificName=scientific_name, SpeciesCode=species_code, AllSpeciesReported=is_complete)

PSSS$RouteIdentifier<-PSSS$Locality
PSSS$BasisOfRecord <- "Observation"
PSSS$CollectionCode <- "PSSS"
PSSS$Continent <-"North America"
PSSS$Country<-"United States"
PSSS$StateProvince<-"Washington"
PSSS$ProtocolType <- "PointCount"
PSSS$ProtocolSpeciesTargeted <- "Waterbirds"
PSSS$ProtocolURL= "https://seattleaudubon.org/wp-content/uploads/2021/01/PSSS_Protocol_2014-15.pdf"
PSSS$SurveyAreaShape = "300 m"
#PSSS$EffortUnit1 = "Party-hours"
PSSS$ObservationDescriptor = "Total Count"
PSSS$ObservationDescriptor2 = "Large flock best estiamte" 

#Now that we have specified all the data columns we can, we will create the BMDE standardized data table. 

#Identify the missing columns of data
BMDE_col<-unique(BMDE$local_name)

missing<-setdiff(BMDE_col, names(PSSS))
PSSS[missing]<-" "
PSSS<-PSSS[BMDE_col]

```

#PSSS Data cleaning

```{r clean}

#add some columns with new names
PSSS<-PSSS %>% mutate(survey_year=YearCollected, survey_month=MonthCollected, survey_day=DayCollected, SiteCode=SurveyAreaIdentifier, latitude=DecimalLatitude, longitude=DecimalLongitude)

#retain desired data columns
dat<-PSSS %>% select(SpeciesCode, survey_year, survey_month, survey_day, SiteCode, SamplingEventIdentifier, SurveyAreaIdentifier, latitude, longitude, DurationInHours, TimeCollected, CollectorNumber, DurationInHours, EffortUnits2, EffortMeasurement2, ObservationCount, ObservationCount2, ObservationCount3, ObservationCount4)

#add doy
dat$survey_year<-as.integer(dat$survey_year)
dat$survey_month<-as.integer(dat$survey_month)
dat$survey_day<-as.integer(dat$survey_day)

dat <- dat %>% mutate(date = ymd(paste(survey_year, survey_month, survey_day, sep = "/")), doy = yday(date))

#drop site with missing lat and long
dat<-dat %>% filter(latitude != "NA")

#assign winter period
dat<-dat %>% mutate(period=ifelse(survey_month %in%c(9, 10, 11, 12), survey_year+1, survey_year))

#Detection Covariates

#create new visibility metric, assign NA the mean value (857)
dat<-dat %>% mutate(visibility=ifelse(EffortMeasurement2=="750m", 750, ifelse(EffortMeasurement2=="500m", 500, ifelse(EffortMeasurement2=="1km", 1000, ifelse(EffortMeasurement2=="100m", 100, ifelse(EffortMeasurement2=="250m", 250, ifelse(EffortMeasurement2=="lt100m", 50, NA))))))) %>% select(-EffortUnits2, -EffortMeasurement2) 

dat<-dat %>% mutate(vis=ifelse(is.na(visibility), 857, visibility)) %>% select(-visibility)

#create a new duration metric, assign NA the mean value (0.71)
#first replace outliers to calculate mean, then replace the outliers with the mean so that we don't have data gaps. 

#duration<-dat %>% filter(DurationInHours<100 & DurationInHours>=0)
#summary(duration) mean(0.71)

dat<-dat %>% mutate(duration=ifelse(DurationInHours<100 & DurationInHours>=0, DurationInHours, 0.71)) %>% select(-DurationInHours)
                    
dat<-dat %>% mutate(duration=replace_na(duration, 0.71)) 

#create a new time of day metric, and assign the NA the mean value (11.19)
#time<-dat %>% filter(TimeCollected<1500 & TimeCollected>=0)
#summary(time) mean(11.19)

dat<-dat %>% mutate(time=ifelse(TimeCollected<1500 & TimeCollected>=400, TimeCollected, 1119)) %>% select(-TimeCollected)
dat<-dat %>% mutate(time=time/100)
dat<-dat %>% mutate(tod=replace_na(time, 11.11)) %>% select(-time)

```

#PSSS Events Matrix 
Need this to properly zero-fill the data matrix

```{r events}

events<-dat %>% select(SiteCode, period, survey_year, survey_month, survey_day, tod, duration, vis, latitude, longitude) %>% distinct()

events<-events %>% mutate(date = ymd(paste(period, survey_month, survey_day, sep = "/")), doy = yday(date))

```

#Limit samples in events matrix

Also, similar to de Zwann, we will want to limit the Only routes that were visited during all four months (12,1,2,3) for at least 3 years within the 20-year study period will be retained. 

```{r month}

#Limit the survey to one per month (sometime extras are run).
events<-events %>% group_by(SiteCode, period, survey_month) %>% slice_max(survey_day) 
events<-events %>% distinct(SiteCode, period, doy, .keep_all = TRUE)

events<-as.data.frame(events)

#filter for desired months
events<-events %>% filter(survey_month %in% c(12,1,2,3))

#site must be visit all 4 times
test<-events %>% group_by(SiteCode, period) %>% summarise(visits=length(survey_month)) %>% filter(visits>=4) %>% select(-visits)

events<-inner_join(events, test, by=c("SiteCode", "period"))

#site must be visited for 3 years
test2<-events %>% group_by(SiteCode) %>% summarise(yrs=n_distinct(period)) %>% filter(yrs>=3) %>% select(-yrs)

events<-inner_join(events, test2, by=c("SiteCode"))

#write.csv(events, "EventsPSSS.csv")
#events<-read.csv("EventsPSSS.csv")

```

#Detection Matrix PSSS
Create the species detection/no detection matrix for target species

Target Species
"BAEA"
"SUSC"
"WWSC"
"BLSC"
"LTDU"

```{r detect}

sp.list<-c("BAEA", "SUSC", "WWSC", "BLSC", "LTDU")

#create sp loop

for(m in 1:length(sp.list)) {
  
 # m<-1 #for testing
  
  sp.dat<-dat %>% filter(SpeciesCode==sp.list[m])
  
  sp.dat<-sp.dat %>% select(period, survey_month, survey_day, SiteCode, ObservationCount)
  
#define the core winter months as Dec-March 
  
  sp.dat<-sp.dat %>% filter(survey_month %in% c(12,1,2,3))
  events1<-events %>% select(-latitude, -longitude)
  
#zero-fill the dataframe
  
  sp.dat<-left_join(events1, sp.dat, by=c("period", "survey_month", "survey_day", "SiteCode"))
  
  sp.dat<-sp.dat %>% distinct(period, survey_month, survey_day, SiteCode, .keep_all = TRUE)
  
sp.dat<-sp.dat %>% mutate(detection=ifelse(ObservationCount>=1, 1, 0))  %>% replace_na(list(detection=0)) %>% mutate(Site=paste(SiteCode, period, sep="-")) %>% select(-ObservationCount, -SiteCode, -period, -survey_day) %>% arrange(Site) 

#recast the data.frame  
  sp<-cast(sp.dat, Site~survey_month, value="detection", fun.aggregate = "max")

#replace Inf in data by NA
 sp <- do.call(data.frame,lapply(sp, function(x) replace(x,is.infinite(x), NA)))
 
 sp<- sp %>% select(X12, X1, X2, X3)
 sp<-as.matrix(sp)
 
 write.csv(sp, paste(sp.list[m], "detectPSSS.csv", sep=""))
 
} # end sp.loop

```

#Observer-level Detection Covariates PSSS

These detection covariates were selected base on modelling outputs from deZwaan's works using the BCCWS (unpublished)

Time of Day = tod
Day of Year = doy
Survey duration = duration
Visibility = vis

```{r detectcov}

cov.dat<-events %>% mutate(Site=paste(SiteCode, period, sep="-"))  %>% select(-SiteCode) %>% arrange(Site)

#Time of Day  
#recast the data.frame  
time<-cast(cov.dat, Site~survey_month, value="tod", fun.aggregate = "max")

#replace Inf in data with the mean value
time <- do.call(data.frame,lapply(time, function(x) replace(x,is.infinite(x), 11.11)))

time<- time %>% select(X12,X1,X2,X3)

#standardize all variables
time<-time %>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

time<-as.matrix(time)
write.csv(time, "detPSSS_tod.csv")

#Duration
#recast the data.frame  
dur<-cast(cov.dat, Site~survey_month, value="duration", fun.aggregate = "max")

#replace Inf in data with the mean value
dur <- do.call(data.frame,lapply(dur, function(x) replace(x,is.infinite(x), 1.7)))

dur<- dur %>% select(X12,X1,X2,X3)

#standardize all variables
dur<-dur %>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

dur<-as.matrix(dur)
write.csv(dur, "detPSSS_dur.csv")

#Visibility
#recast the data.frame  
vis<-cast(cov.dat, Site~survey_month, value="vis", fun.aggregate = "max")

#replace Inf in data with the mean value
vis <- do.call(data.frame,lapply(vis, function(x) replace(x,is.infinite(x), 885)))

vis<- vis %>% select(X12,X1,X2,X3)

#standardize all variables
vis<-vis %>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

vis<-as.matrix(vis)
write.csv(vis, "detPSSS_vis.csv")

#Day of year
#doy needs to start at 1 and sequence upwards. Import data specific file to do this
doy<-read.csv("doy.csv")
doy<-left_join(cov.dat, doy, by="doy")

#recast the data.frame  
doy<-cast(doy, Site~survey_month, value="Day", fun.aggregate = "max")

#replace Inf in data with the mean value
doy<- do.call(data.frame,lapply(doy, function(x) replace(x,is.infinite(x), 57.07)))
doy<- doy %>% select(X12,X1,X2,X3)

#standardize all variables
doy<-doy%>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

doy<-as.matrix(doy)
write.csv(vis, "detPSSS_doy.csv")

```

#Site-level Occupancy Covariates

Occupancy covaraites of interest include:
latitude 
longitude
year
collection
region** (still need to assign)

```{r occcov}

occ.dat<-events %>% select(SiteCode, period, doy, latitude, longitude) %>% distinct(period, SiteCode, .keep_all = TRUE)

#retain only those in the events layer
#occ.dat<-left_join(events, occ.dat, by=c("SiteCode", "period", "doy"))

occ.dat<-occ.dat %>% mutate(Site=paste(SiteCode, period, sep="-")) %>% arrange(Site)  %>% distinct(Site, .keep_all = TRUE) #%>% select(-SiteCode)

#standardize year so that the first year of the analysis is 1
min.yr<-min(occ.dat$period)
max.yr<-max(occ.dat$period)

occ.dat<-occ.dat %>% mutate(year_st=(period-min.yr)+1)

#standardize all variables
#need to do this in the next step
#occ.dat<-occ.dat %>% mutate_at(c('latitude', 'longitude'), ~(scale(.) %>% as.vector)) %>% select(Site, year_st, latitude, longitude)

#Assign Collection identifier
occ.dat$collection<-2

write.csv(occ.dat, "obsCovsPSSS.csv")
  
```


#Review Edits

#Detection matrix all sea ducks combined
Create the species detection/no detection matrix for target species


```{r detect}

dat<-dat %>% mutate(SpeciesCode = ifelse(SpeciesCode %in% c("SUSC", "WWSC", "BLSC", "LTDU"), "ALL", SpeciesCode))

sp.list<-c("BAEA", "ALL")

#create sp loop

for(m in 1:length(sp.list)) {
  
 # m<-1 #for testing
  
  sp.dat<-dat %>% filter(SpeciesCode==sp.list[m])
  
  sp.dat<-sp.dat %>% select(period, survey_month, survey_day, SiteCode, ObservationCount)
  
#define the core winter months as Dec-March 
  
  sp.dat<-sp.dat %>% filter(survey_month %in% c(12,1,2,3))
  events1<-events %>% select(-latitude, -longitude)
  
#zero-fill the dataframe
  
  sp.dat<-left_join(events1, sp.dat, by=c("period", "survey_month", "survey_day", "SiteCode"))
  
  sp.dat<-sp.dat %>% distinct(period, survey_month, survey_day, SiteCode, .keep_all = TRUE)
  
sp.dat<-sp.dat %>% mutate(detection=ifelse(ObservationCount>=1, 1, 0))  %>% replace_na(list(detection=0)) %>% mutate(Site=paste(SiteCode, period, sep="-")) %>% select(-ObservationCount, -SiteCode, -period, -survey_day) %>% arrange(Site) 

#recast the data.frame  
  sp<-cast(sp.dat, Site~survey_month, value="detection", fun.aggregate = "max")

#replace Inf in data by NA
 sp <- do.call(data.frame,lapply(sp, function(x) replace(x,is.infinite(x), NA)))
 
 sp<- sp %>% select(X12, X1, X2, X3)
 sp<-as.matrix(sp)
 
 write.csv(sp, paste(sp.list[m], "detectPSSS2.csv", sep=""))
 
} # end sp.loop

```

#Observer-level Detection Covariates PSSS

These detection covariates were selected base on modelling outputs from deZwaan's works using the BCCWS (unpublished)

Time of Day = tod
Day of Year = doy
Survey duration = duration
Visibility = vis

```{r detectcov}

cov.dat<-events %>% mutate(Site=paste(SiteCode, period, sep="-"))  %>% select(-SiteCode) %>% arrange(Site)

#Time of Day  
#recast the data.frame  
time<-cast(cov.dat, Site~survey_month, value="tod", fun.aggregate = "max")

#replace Inf in data with the mean value
time <- do.call(data.frame,lapply(time, function(x) replace(x,is.infinite(x), 11.11)))

time<- time %>% select(X12,X1,X2,X3)

#standardize all variables
time<-time %>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

time<-as.matrix(time)
write.csv(time, "detPSSS_tod2.csv")

#Duration
#recast the data.frame  
dur<-cast(cov.dat, Site~survey_month, value="duration", fun.aggregate = "max")

#replace Inf in data with the mean value
dur <- do.call(data.frame,lapply(dur, function(x) replace(x,is.infinite(x), 1.7)))

dur<- dur %>% select(X12,X1,X2,X3)

#standardize all variables
dur<-dur %>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

dur<-as.matrix(dur)
write.csv(dur, "detPSSS_dur2.csv")

#Visibility
#recast the data.frame  
vis<-cast(cov.dat, Site~survey_month, value="vis", fun.aggregate = "max")

#replace Inf in data with the mean value
vis <- do.call(data.frame,lapply(vis, function(x) replace(x,is.infinite(x), 885)))

vis<- vis %>% select(X12,X1,X2,X3)

#standardize all variables
vis<-vis %>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

vis<-as.matrix(vis)
write.csv(vis, "detPSSS_vis2.csv")

#Day of year
#doy needs to start at 1 and sequence upwards. Import data specific file to do this
doy<-read.csv("doy.csv")
doy<-left_join(cov.dat, doy, by="doy")

#recast the data.frame  
doy<-cast(doy, Site~survey_month, value="Day", fun.aggregate = "max")

#replace Inf in data with the mean value
doy<- do.call(data.frame,lapply(doy, function(x) replace(x,is.infinite(x), 57.07)))
doy<- doy %>% select(X12,X1,X2,X3)

#standardize all variables
doy<-doy%>% mutate_at(c('X12', 'X1', 'X2', 'X3'), ~(scale(.) %>% as.vector))

doy<-as.matrix(doy)
write.csv(vis, "detPSSS_doy2.csv")

```
