==> 03-MultiOcc_Anal.Rmd <==
---
title: "MultiOcc_Anal"
author: "Danielle Ethier"
date: '2022-07-06'
output: html_document
editor_options: 
  chunk_output_type: console
---

Now that we have build the multi-species occupancy model framework, we can test the analysis. 

#Load libraries

```{r libraries}

#You can install naturecounts from GitHub with the remotes package if you don't have it already. 
#install.packages("remotes")
#remotes::install_github("BirdsCanada/naturecounts")

require(naturecounts)
require(tidyverse)
require(unmarked)
require(lubridate)
require(reshape)
require(reshape2)
require(nloptr)
library(beepr)
require(ggpubr)
library(DMwR2)

#library(rgdal)
library(sf)
library(ggmap)
library(leaflet)
library(mapview)
library(rnaturalearth)

library("ggmap") 
register_stadiamaps("1cdeb586-3f5a-4ef9-813b-02375efa9f21") 

out.dir <- paste("Output/")

```

#Format the data for the unmarked framework

```{r unmarked}

#species data BCCWS
BAEA1<-read.csv("BAEAdetectBCCWS.csv", row.names="X")
SUSC1<-read.csv("SUSCdetectBCCWS.csv", row.names="X")
WWSC1<-read.csv("WWSCdetectBCCWS.csv", row.names="X")
BLSC1<-read.csv("BLSCdetectBCCWS.csv", row.names="X")
LTDU1<-read.csv("LTDUdetectBCCWS.csv", row.names="X")

#Detection level covariates BCCWS
doy1<-read.csv("detBCCWS_doy.csv", row.name="X")
tod1<-read.csv("detBCCWS_tod.csv", row.name="X")
vis1<-read.csv("detBCCWS_vis.csv", row.name="X")
dur1<-read.csv("detBCCWS_dur.csv", row.name="X")

#species data PSSS
BAEA2<-read.csv("BAEAdetectPSSS.csv", row.names="X")
SUSC2<-read.csv("SUSCdetectPSSS.csv", row.names="X")
WWSC2<-read.csv("WWSCdetectPSSS.csv", row.names="X")
BLSC2<-read.csv("BLSCdetectPSSS.csv", row.names="X")
LTDU2<-read.csv("LTDUdetectPSSS.csv", row.names="X")

#Detection level covariates PSSS
doy2<-read.csv("detPSSS_doy.csv", row.name="X")
tod2<-read.csv("detPSSS_tod.csv", row.name="X")
vis2<-read.csv("detPSSS_vis.csv", row.name="X")
dur2<-read.csv("detPSSS_dur.csv", row.name="X")

BAEA<-rbind(BAEA1, BAEA2)
SUSC<-rbind(SUSC1, SUSC2)
WWSC<-rbind(WWSC1, WWSC2)
BLSC<-rbind(BLSC1, BLSC2)
LTDU<-rbind(LTDU1, LTDU2)

y_list<-list(BAEA=as.matrix(BAEA), 
             SUSC=as.matrix(SUSC), 
             WWSC=as.matrix(WWSC), 
             BLSC=as.matrix(BLSC), 
             LTDU=as.matrix(LTDU))

doy<-rbind(doy1, doy2)
tod<-rbind(tod1, tod2)
vis<-rbind(vis1,vis2)
dur<-rbind(dur1, dur2)

det_list<-list(TimeofDay=tod, DayofYear = doy, Visibility=vis, Duration = dur)

#Observation level covaraites
occBCCWS_covs<-read.csv("obsCovsBCCWS.csv", row.names="X")
occPSSS_covs<-read.csv("obsCovsPSSS.csv", row.names="X")

occ_covs<-rbind(occBCCWS_covs, occPSSS_covs)

plot_data<-occ_covs %>% select(SiteCode, period, latitude, longitude)

#Standardize Lat but first get values you need to back transform later
latmin = min(occ_covs$latitude)
latmax = max(occ_covs$latitude)
latmean = mean(occ_covs$latitude)
latsd = sd(occ_covs$latitude)

#Standardize Lat but first get values you need to back transform later
longmin = min(occ_covs$longitude)
longmax = max(occ_covs$longitude)
longmean = mean(occ_covs$longitude)
longsd = sd(occ_covs$longitude)

lat.s<-scale(occ_covs$latitude)
long.s<-scale(occ_covs$longitude)

occ_covs<-occ_covs %>% mutate_at(c('latitude', 'longitude'), ~(scale(.) %>% as.vector)) %>% select(Site, SiteCode, year_st, latitude, longitude)

yr.min<-2009
yr.max<-2022

```

Make a map
```{r plot}

#Create spatial data
plot <- plot_data %>% select("longitude", "latitude")%>% distinct()
plot_sf <- st_as_sf(plot, coords = c("longitude", "latitude"), crs = 4326)

#Create the base layer map
map <- get_stadiamap(bbox = as.numeric(st_bbox(plot_sf)), zoom = 5)

#Create a new plot for each year
plot<-ggmap(map) + 
  geom_point(data = plot_data, aes(x = longitude, y = latitude))
print(plot)

```

#Create the occupancy model framework and get summary outputs

```{r occ sum}

msom_data<-unmarkedFrameOccuMulti(y=y_list, siteCovs=occ_covs, obsCovs=det_list)

summary(msom_data)

plot(msom_data)

```


#Fit a model with no covariates, assuming independence

To start, we assume independence among species. We do this by only allowing 1st order natural parameters (maxOrder = 1). This is the same as fitting 5 single-species occupancy models

```{r fit1}

#fit model with no covariates to test 
summary(msom_data) # to get summary of dataset. Useful output for results of MS. 

fit_1 <- occuMulti(detformulas = c('~1', '~1', '~1', '~1', '~1'),
                   stateformulas = c('~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 1,
                   data = msom_data)
summary(fit_1)

#AIC = 37925.5


```

#Fit a model with no covarites (intercept-only), assuming dependence

```{r fit2}

fit_2<- occuMulti(detformulas = c('~1', '~1', '~1', '~1', '~1'),
                   stateformulas = c('~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_2)

#AIC = 36881, suggesting that there is strong support for the model incorporating dependence.



```

#Fit a model covarites for duration only
 
```{r fitD}

fit_D<- occuMulti(detformulas = c('~ Duration', '~ Duration', '~ Duration', '~  Duration', '~ Duration'),
                   stateformulas = c('~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_D)


#AIC = 35183.07 

#get some summary outputs



```

#Fit a model covarites for visibility

```{r fitV}

fit_V<- occuMulti(detformulas = c('~ Visibility', '~ Visibility', '~ Visibility', '~ Visibility', '~ Visibility'),
                   stateformulas = c('~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_V)


#AIC = 35200.2

#get some summary outputs



```

#Fit a model covarites for duration and visibility
**Top Detection Model**

```{r fitDV}

fit_DV<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                   stateformulas = c('~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_DV)


#AIC = 35163.1 

```


#Fit a model covarites for maginal model (year only)


```{r fit year}

fit_M0<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                   stateformulas = c('~year_st', '~year_st', '~year_st', '~year_st', '~year_st', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_M0)


#AIC = 35112.92

#get some summary outputs



```

#Fit a model covarites for maginal model (with year+lat and lat)


```{r fit lat}

fit_lat<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                   stateformulas = c('~year_st+latitude', '~year_st+latitude', '~year_st+latitude', '~year_st+latitude', '~year_st+latitude', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_lat)


#AIC = 34482.92

fit_lat0<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                   stateformulas = c('~latitude', '~latitude', '~latitude', '~latitude', '~latitude', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_lat0)

#AIC: 

```


#Fit a model covarites for maginal model (with year+long and long)

```{r fit long}

fit_long<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                   stateformulas = c('~year_st+longitude', '~year_st+longitude', '~year_st+longitude', '~year_st+longitude', '~year_st+longitude', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_long)

beep(4)

#AIC = 34501.09 

#get some summary outputs


fit_long0<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                   stateformulas = c('~longitude', '~longitude', '~longitude', '~longitude', '~longitude', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_long0)


```

#Fit a model full model 
**Best Model**

Detection covaraites fixed as visibility and duration (based on de Zwann)

Interested in the effect of year on occupancy probability, while taking into account spatial autocorrelation using lat and long. 

```{r fit4}

fit_M1<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                   stateformulas = c('~year_st+latitude+longitude', '~year_st+latitude+longitude', '~year_st+latitude+longitude', '~year_st+latitude+longitude', '~year_st+latitude+longitude', '~1', '~1', '~1', '~1', '~1','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_M1)

#AIC = 36454.16

#get some summary outputs of top model

Out_det<-summary(fit_M1)$det
write.csv(Out_det, "Output/Out_det.csv")
Out_state<-summary(fit_M1)$state
write.csv(Out_state, "Output/Out_state.csv")

```

#Fit second order models (with year and lat and long)

```{r fit test}

fit_test<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                     stateformulas = c('~year_st', '~year_st', '~year_st', '~year_st', '~year_st', '~latitude', '~latitude', '~latitude', '~latitude', '~latitude','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_test)

beep(4)

#AIC = 34826

fit_test2<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                     stateformulas = c('~year_st', '~year_st', '~year_st', '~year_st', '~year_st', '~longitude', '~longitude', '~longitude', '~longitude', '~longitude','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_test2)

beep(4)

#AIC = 34669

fit_test2.1<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                     stateformulas = c('~year_st', '~year_st', '~year_st', '~year_st', '~year_st', '~longitude+latitude', '~longitude+latitude', '~longitude+latitude', '~longitude+latitude', '~longitude+latitude','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_test2.1)

beep(4)

#AIC = 34669

fit_test3<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                     stateformulas = c('~1', '~1', '~1', '~1', '~1', '~longitude', '~longitude', '~longitude', '~longitude', '~longitude','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_test3)

#AIC = 34723

fit_test4<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                     stateformulas = c('~1', '~1', '~1', '~1', '~1','~latitude', '~latitude', '~latitude', '~latitude', '~latitude','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_test4)

fit_test4.1<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                     stateformulas = c('~1', '~1', '~1', '~1', '~1','~latitude+longitude', '~latitude+longitude', '~latitude+longitude', '~latitude+longitude', '~latitude','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_test4.1)

#AIC = 34875

fit_test5<- occuMulti(detformulas = c('~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration', '~ Visibility + Duration'),
                     stateformulas = c('~1', '~1', '~1', '~1', '~1', '~year_st', '~year_st', '~year_st', '~year_st', '~year_st','~1', '~1', '~1', '~1', '~1'),
                   maxOrder = 2,
                   data = msom_data)
summary(fit_test5)

beep(4)

```

#Model Selection
```{r model selection}

mods1 <- fitList(fit_1, fit_2)
modSel(mods1)
                
mods2 <- fitList(fit_D, fit_V, fit_DV) 
modSel(mods2)
                
mods3 <- fitList(fit_M0, fit_lat0, fit_lat, fit_long0, fit_long)
modSel(mods3)   

mods<-fitList(fit_M1)
modSel(mods)
             
mods4 <- fitList(fit_test3, fit_test4, fit_test4.1, fit_test, fit_test2, fit_test2.1, fit_test5)
modSel(mods4)
```

**Top model is M1**

#M1:Marginal occupancy probability over time and latitude

Omit the `cond` argument if you wish to calculate marginal occupancy probability
Averages over presence / absence of all other species


```{r predict marginal time}

sp.list<-c("BAEA", "SUSC", "WWSC", "BLSC", "LTDU")

time_marg <- as.data.frame(matrix(data = NA, nrow = 1, ncol = 6, byrow = FALSE, dimnames = NULL))

names(time_marg) <- c("hd", "occupancy", "low", "high", "Year", "Species")

#only need to create the table once per analysis   

write.table(time_marg, file = paste(out.dir,  "time_marg.csv"), row.names = FALSE, append = FALSE, quote = FALSE, sep = ",")
    

#Margial effect of year  
nd_marg <- data.frame(
 year_st = seq(min(occ_covs$year_st), max(occ_covs$year_st),length.out = 100),
 longitude = rep(mean(occ_covs$longitude), 100), 
 latitude = rep(mean(occ_covs$latitude), 100)
)

#sp loop
for(m in 1:length(sp.list)) {
  
 # m<-1 #for testing
  
marg <- predict(fit_M1, type = 'state', species = sp.list[m], newdata = nd_marg)

gg_df_marg <- data.frame(
  hd = nd_marg$year_st,
  occupancy = marg$Predicted,
  low = marg$lower,
  high = marg$upper
)

#Revert year_st back to Year
gg_df_marg<-gg_df_marg %>% mutate(Year=(hd-1)+yr.min)   
gg_df_marg$Species<-sp.list[m]

marg_fig <- ggplot(gg_df_marg, aes(x = Year, y = occupancy)) +
  geom_ribbon(aes(ymin = low, ymax = high), alpha = 0.5) +
  geom_line() +
  ylab('') +
  xlab('') +
  ggtitle(sp.list[m])+
  theme_classic()+
 # scale_fill_brewer(palette="Blues")+
  scale_x_continuous(limits = c(2010, 2022), breaks = c(2010, 2012, 2014, 2016, 2018, 2020, 2022))+
  scale_y_continuous(limits = c(0, 1))+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  theme(text = element_text(size = 30))   

marg_fig

ggsave(paste(out.dir, sp.list[m], "_marginal_time.jpeg", sep=""), width = 11, height = 8, device = "jpeg")
while (!is.null(dev.list()))  dev.off()

write.table(gg_df_marg, file = paste(out.dir, "time_marg.csv"), row.names = FALSE, append = TRUE, 	quote = FALSE, sep = ",", col.names = FALSE)

} #end sp Year loop

marg_time<-read.csv("Output/ time_marg.csv")
marg_time<-na.omit(marg_time)

ggplot(marg_time, aes(x = Year, y = occupancy)) +
  geom_ribbon(aes(ymin = low, ymax = high), alpha = 0.5) +
  geom_line() +
  ylab('Marginal occupancy') +
  xlab('Year') +
  facet_wrap(~Species, scales="free")+
  theme_classic()+
  scale_x_continuous(limits = c(2010, 2022), breaks = c(2010, 2012, 2014, 2016, 2018, 2020, 2022))+
  scale_y_continuous(limits = c(0, 1))+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  theme(text = element_text(size = 20)) 


```


```{r predict marginal latitude}

sp.list<-c("BAEA", "SUSC", "WWSC", "BLSC", "LTDU")

#Margial effect of lat
nd_marg <- data.frame(
  year_st = rep(mean(occ_covs$year_st), 100),
  longitude = rep(mean(occ_covs$longitude), 100),
  latitude = seq(min(occ_covs$latitude), max(occ_covs$latitude),length.out = 100))

lat_marg <- as.data.frame(matrix(data = NA, nrow = 1, ncol = 6, byrow = FALSE, dimnames = NULL))

names(lat_marg) <- c("hd", "occupancy", "low", "high",  "Lat", "Species")

#only need to create the table once per analysis   

write.table(lat_marg, file = paste(out.dir,  "lat_marg.csv"), row.names = FALSE, append = FALSE, quote = FALSE, sep = ",")
    

#sp loop
for(m in 1:length(sp.list)) {
  
 # m<-1 #for testing
  
marg <- predict(fit_M1, type = 'state', species = sp.list[m], newdata = nd_marg)

gg_df_marg <- data.frame(
  hd = nd_marg$latitude,
  occupancy = marg$Predicted,
  low = marg$lower,
  high = marg$upper
)

#Revert latitude back to natural scale 
#gg_df_marg<-gg_df_marg %>% mutate(Lat=hd*(latmax - latmin) + latmin)
gg_df_marg$Lat <- gg_df_marg$hd* attr(lat.s, 'scaled:scale') + attr(lat.s, 'scaled:center')

gg_df_marg$Species<-sp.list[m]

marg_fig <- ggplot(gg_df_marg, aes(x = Lat, y = occupancy)) +
  geom_ribbon(aes(ymin = low, ymax = high), alpha = 0.5) +
  geom_line() +
  ylab('') +
  xlab('') +
  ggtitle(sp.list[m])+
  theme_classic()+
  theme(text = element_text(size = 25))

marg_fig

ggsave(paste(out.dir, sp.list[m], "_marginal_latitude.jpeg", sep=""), width = 11, height = 8, device = "jpeg")
while (!is.null(dev.list()))  dev.off()

write.table(gg_df_marg, file = paste(out.dir, "lat_marg.csv"), row.names = FALSE, append = TRUE, 	quote = FALSE, sep = ",", col.names = FALSE)

} #end sp Year loop

marg_lat<-read.csv("Output/ lat_marg.csv")
marg_lat<-na.omit(marg_lat)

ggplot(marg_lat, aes(x = Lat, y = occupancy)) +
  geom_ribbon(aes(ymin = low, ymax = high), alpha = 0.5) +
  geom_line() +
  ylab('Marginal occupancy') +
  xlab('Latitude') +
  facet_wrap(~Species, scales="free")+
  theme_classic()+
   theme(text = element_text(size = 20)) 

```


```{r predict marginal longitude}

sp.list<-c("BAEA", "SUSC", "WWSC", "BLSC", "LTDU")

long_marg <- as.data.frame(matrix(data = NA, nrow = 1, ncol = 6, byrow = FALSE, dimnames = NULL))

names(long_marg) <- c("hd", "occupancy", "low", "high",  "Long", "Species")

#only need to create the table once per analysis   

write.table(long_marg, file = paste(out.dir,  "long_marg.csv"), row.names = FALSE, append = FALSE, quote = FALSE, sep = ",")

#Margial effect of long
nd_marg <- data.frame(
  year_st = rep(mean(occ_covs$year_st), 100),
  latitude = rep(mean(occ_covs$latitude), 100),
  longitude = seq(min(occ_covs$longitude), max(occ_covs$longitude),length.out = 100))

#sp loop
for(m in 1:length(sp.list)) {
  
 # m<-1 #for testing
  
marg <- predict(fit_M1, type = 'state', species = sp.list[m], newdata = nd_marg)

gg_df_marg <- data.frame(
  hd = nd_marg$longitude,
  occupancy = marg$Predicted,
  low = marg$lower,
  high = marg$upper
)

#Revert latitude back to natural scale 
#gg_df_marg<-gg_df_marg %>% mutate(Long=hd*(longmax - longmin) + longmin)

gg_df_marg$Long <- gg_df_marg$hd* attr(long.s, 'scaled:scale') + attr(long.s, 'scaled:center')

gg_df_marg$Species<-sp.list[m]

marg_fig <- ggplot(gg_df_marg, aes(x = Long, y = occupancy)) +
  geom_ribbon(aes(ymin = low, ymax = high), alpha = 0.5) +
  geom_line() +
  ylab('') +
  xlab('') +
  ggtitle(sp.list[m])+
  theme_classic()+
  theme(text = element_text(size = 25))

marg_fig

ggsave(paste(out.dir, sp.list[m], "_marginal_longitude.jpeg", sep=""), width = 11, height = 8, device = "jpeg")
while (!is.null(dev.list()))  dev.off()

write.table(gg_df_marg, file = paste(out.dir, "long_marg.csv"), row.names = FALSE, append = TRUE, 	quote = FALSE, sep = ",", col.names = FALSE)

} #end sp Year loop

marg_long<-read.csv("Output/ long_marg.csv")
marg_long<-na.omit(marg_long)

ggplot(marg_long, aes(x = Long, y = occupancy)) +
  geom_ribbon(aes(ymin = low, ymax = high), alpha = 0.5) +
  geom_line() +
  ylab('Marginal occupancy') +
  xlab('Longitude') +
  facet_wrap(~Species, scales="free")+
  theme_classic()+
   scale_x_continuous(limits = c(-125.25, -122.25), breaks = c(-125, -124, -123, -122))+
   theme(text = element_text(size = 20)) 



```

#M1: Conditional occupancy probability

Calculation of conditional and marginal occupancy probabilities is done with the predict function.

Create a data.frame for predictions
The procedure is equivalent to creating data frames for all other applications of predict

Include complete range of observed *year_st*; hold all other variables at their mean. 
Predicting prey species occurrence when bald eagles are present

`Species` indicates which species we assume when predicting occupancy
`cond` indicates which species we are assuming is present or absent

#Year
```{r predict year}

#sp list excluding BAEA
sp.list2<-c("SUSC", "WWSC", "BLSC", "LTDU")

yr_cond <- data.frame(
  latitude = rep(mean(occ_covs$latitude), 100),
  year_st = seq(min(occ_covs$year_st), max(occ_covs$year_st),length.out = 100))

#sp loop
for(n in 1:length(sp.list2)) {
  
BAEA_mod <- predict(fit_M1, type = 'state', species = sp.list2[n], cond = 'BAEA', newdata = yr_cond)

#putting a `-`in front of BAEA tells predict you wish to assume they are absent

BAEA_mod_0 <- predict(fit_M1, type = 'state', species = sp.list2[n], cond = '-BAEA', newdata = yr_cond)

#formatting ggplot

gg_df_cond <- data.frame(
  year_st = rep(yr_cond$year_st, 2),
  occupancy = c(BAEA_mod$Predicted,
                BAEA_mod_0$Predicted),
  low = c(BAEA_mod$lower,
          BAEA_mod_0$lower),
  high = c(BAEA_mod$upper,
           BAEA_mod_0$upper),
  conditional = rep(c('BAEA Present', 'BAEA Absent'),each = 100)
)

#Revert year_st back to Year
gg_df_cond<-gg_df_cond %>% mutate(Year=(year_st+1)+yr.min)

cond_fig <- ggplot(gg_df_cond, aes(x = Year, y = occupancy,
                                   group = conditional)) +
  geom_ribbon(aes(ymin = low, ymax = high, fill = conditional)) +
  geom_line() +
  ylab('') +
  xlab('') +
  labs(fill = '') +
  theme_classic()+
  ggtitle(sp.list2[n])+
  scale_fill_brewer(palette="Blues")+
  scale_x_continuous(limits = c(2010, 2022), breaks = c(2010, 2012, 2014, 2016, 2018, 2020, 2022))+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  theme(text = element_text(size = 30)) +
  theme(legend.position = "none")    
  
cond_fig

ggsave(paste(out.dir, sp.list2[n], "_conditional_time.jpeg", sep=""), width = 11, height = 8, device = "jpeg")
while (!is.null(dev.list()))  dev.off()

} # end sp con loop

beep(4)

```


#Latitude
```{r predict latitude}

#sp list excluding BAEA
sp.list2<-c("SUSC", "WWSC", "BLSC", "LTDU")

lat_cond <- data.frame(
  year_st = rep(mean(occ_covs$year_st), 100),
  latitude = seq(min(occ_covs$latitude), max(occ_covs$latitude),length.out = 100))

#sp loop
for(n in 1:length(sp.list2)) {
  
BAEA_mod <- predict(fit_M1, type = 'state', species = sp.list2[n], cond = 'BAEA', newdata = lat_cond)

#putting a `-`in front of BAEA tells predict you wish to assume they are absent

BAEA_mod_0 <- predict(fit_M1, type = 'state', species = sp.list2[n], cond = '-BAEA', newdata = lat_cond)

#Revert latitude back to natural scale 
lat_cond<-lat_cond %>% mutate(Lat=latitude*(latmax - latmin) + latmin)

#formatting ggplot

gg_df_cond <- data.frame(
  latitude = rep(lat_cond$Lat, 2),
  occupancy = c(BAEA_mod$Predicted,
                BAEA_mod_0$Predicted),
  low = c(BAEA_mod$lower,
          BAEA_mod_0$lower),
  high = c(BAEA_mod$upper,
           BAEA_mod_0$upper),
  conditional = rep(c('BAEA Present', 'BAEA Absent'),each = 100)
)

cond_fig <- ggplot(gg_df_cond, aes(x = latitude, y = occupancy,
                                   group = conditional)) +
  geom_ribbon(aes(ymin = low, ymax = high, fill = conditional)) +
  geom_line() +
  ylab('') +
  xlab('') +
  labs(fill = '') +
   ggtitle(sp.list2[n])+
  scale_fill_brewer(palette="Blues")+
  theme_classic()+
  theme(text = element_text(size = 30))+
  theme(legend.position = "none")

cond_fig

ggsave(paste(out.dir, sp.list2[n], "_conditional_latitude.jpeg", sep=""), width = 11, height = 8, device = "jpeg")
while (!is.null(dev.list()))  dev.off()


} #end species loop

beep(4)
```

#Parameter Estimates Top Model

```{r parameters}

## Bootstrapping confidence intervals with 100 samples

#Detection

BAEA_det<-predict(fit_M1, 'det', species=1) #BAEA
BAEA_det<-BAEA_det %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
BAEA_det<-BAEA_det %>% select(mean, se, LCI, UCI) %>% distinct()

SUSC_det<-predict(fit_M1, 'det', species=2)#SUSC
SUSC_det<-SUSC_det %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
SUSC_det<-SUSC_det %>% select(mean, se, LCI, UCI) %>% distinct()

WWSC_det<-predict(fit_M1, 'det', species=3)#WWSC
WWSC_det<-WWSC_det %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
WWSC_det<-WWSC_det %>% select(mean, se, LCI, UCI) %>% distinct()

BLSC_det<-predict(fit_M1, 'det', species=4)#BLSC
BLSC_det<-BLSC_det %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
BLSC_det<-BLSC_det %>% select(mean, se, LCI, UCI) %>% distinct()


LTDU_det<-predict(fit_M1, 'det', species=5)#LTDU
LTDU_det<-LTDU_det %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
LTDU_det<-LTDU_det %>% select(mean, se, LCI, UCI) %>% distinct()

#Combine and plot
all_det<-rbind(BAEA_det, SUSC_det, WWSC_det, BLSC_det, LTDU_det)
all_det$Species<-c("BAEA", "SUSC", "WWSC", "BLSC", "LTDU")

ggplot(data=all_det, aes(x=Species, y=mean))+
  geom_errorbar(aes(ymin=LCI, ymax=UCI), width=.1) +
  geom_line()+
  geom_point()+
  theme_classic()+
  ylab("Detection probability") +
  theme(text = element_text(size = 20))

# marginal occupancy
BAEA<-predict(fit_M1,'state',species=1) #BAEA
BAEA<-BAEA %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
BAEA<-BAEA %>% select(mean, se, LCI, UCI) %>% distinct()

SUSC<-predict(fit_M1,'state',species=2) #SUSC
SUSC<-SUSC %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
SUSC<-SUSC %>% select(mean, se, LCI, UCI) %>% distinct()

WWSC<-predict(fit_M1,'state',species=3) #WWSC
WWSC<-WWSC %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
WWSC<-WWSC %>% select(mean, se, LCI, UCI) %>% distinct()

BLSC<-predict(fit_M1,'state',species=4) #BLSC
BLSC<-BLSC %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
BLSC<-BLSC %>% select(mean, se, LCI, UCI) %>% distinct()

LTDU<-predict(fit_M1,'state',species=5) #LTDU
LTDU<-LTDU %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
LTDU<-LTDU %>% select(mean, se, LCI, UCI) %>% distinct()

#Combine and plot
all_marg<-rbind(BAEA, SUSC, WWSC, BLSC, LTDU)
all_marg$Species<-c("BAEA", "SUSC", "WWSC", "BLSC", "LTDU")

ggplot(data=all_marg, aes(x=Species, y=mean))+
  geom_errorbar(aes(ymin=LCI, ymax=UCI), width=.1) +
  geom_line()+
  geom_point()+
  theme_classic()+
  ylab("Marginal occupancy probability") +
  xlab("")+
  theme(text = element_text(size = 20))
  
# conditional occupancy given bald eagles are present
SUSC_P<-predict(fit_M1,'state',species=2,cond='BAEA') #SUSC | BAEA
SUSC_P<-SUSC_P %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
SUSC_P<-SUSC_P %>% select(mean, se, LCI, UCI) %>% distinct()

WWSC_P<-predict(fit_M1,'state',species=3,cond='BAEA') #WWSC | BAEA
WWSC_P<-WWSC_P %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
WWSC_P<-WWSC_P %>% select(mean, se, LCI, UCI) %>% distinct()

BLSC_P<-predict(fit_M1,'state',species=4,cond='BAEA') #BLSC | BAEA
BLSC_P<-BLSC_P %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
BLSC_P<-BLSC_P %>% select(mean, se, LCI, UCI) %>% distinct()

LTDU_P<-predict(fit_M1,'state',species=5,cond='BAEA') #LTDU | BAEA
LTDU_P<-LTDU_P %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
LTDU_P<-LTDU_P %>% select(mean, se, LCI, UCI) %>% distinct()

# conditional occupancy given bald eagles are absent
SUSC_A<-predict(fit_M1,'state',species=2,cond='-BAEA') #SUSC | -BAEA
SUSC_A<-SUSC_A %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
SUSC_A<-SUSC_A %>% select(mean, se, LCI, UCI) %>% distinct()

WWSC_A<-predict(fit_M1,'state',species=3,cond='-BAEA') #WWSC | -BAEA
WWSC_A<-WWSC_A %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
WWSC_A<-WWSC_A %>% select(mean, se, LCI, UCI) %>% distinct()

BLSC_A<-predict(fit_M1,'state',species=4,cond='-BAEA') #BLSC | -BAEA
BLSC_A<-BLSC_A %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
BLSC_A<-BLSC_A %>% select(mean, se, LCI, UCI) %>% distinct()

LTDU_A<-predict(fit_M1,'state',species=5,cond='-BAEA') #LTDU | -BAEA
LTDU_A<-LTDU_A %>% mutate(mean= mean(Predicted), se=mean(SE), LCI=mean(lower), UCI = mean(upper))
LTDU_A<-LTDU_A %>% select(mean, se, LCI, UCI) %>% distinct()

#Combine and plot
all_cond<-rbind(SUSC_P, WWSC_P, BLSC_P, LTDU_P, SUSC_A, WWSC_A, BLSC_A, LTDU_A)
all_cond$Species<-c("SUSC", "WWSC", "BLSC", "LTDU", "SUSC", "WWSC", "BLSC", "LTDU")
all_cond$BaldEagle<-c("Present", "Present", "Present", "Present", "Absent", "Absent", "Absent", "Absent")

dodge <- position_dodge(width=0.5) 

ggplot(data=new_cond, aes(x=Species, y=mean, color=BaldEagle))+
  geom_errorbar(aes(ymin=LCI, ymax=UCI), width=.1, position=dodge) +
  geom_line(position=dodge)+
  geom_point(position=dodge)+
  theme_classic()+
  labs(fill="Bald Eagle")+
  ylab("Conditional occupancy probability") +
  xlab("")+
  scale_colour_manual(values = c('black','darkgrey')) +
  theme(text = element_text(size = 20))


#Significant test
SUSC_P<-predict(fit_M1,'state',species=2,cond='BAEA') #SUSC | BAEA
SUSC_A<-predict(fit_M1,'state',species=2,cond='-BAEA') #SUSC | -BAEA

t.test(SUSC_P$Predicted, SUSC_A$Predicted)
t.test(WWSC_P$Predicted, WWSC_A$Predicted)
t.test(BLSC_P$Predicted, BLSC_A$Predicted)
t.test(LTDU_P$Predicted, LTDU_A$Predicted)


```


bt<-parboot((fit_M1, nsim=30))
unmarkedFitOccuMulti object describing the model fit.
